updated comment section
This edited version was to accommodate the presence of identical SP1-SP2 index combination. Previous version grabs the sample name from the index file to preserve order, but due to the identical indices, the copy-pasting script now looks for the sample name inside the log files. This also means order of sample names is no longer preserved (there is another .py script that does this, may incorporate at some point).